rec: stop supporting ucontext flavor for context switching
