+++ pdns/recursordist/configure.ac
@@ -52,18 +52,7 @@ AC_FUNC_STRERROR_R
 PDNS_CHECK_CLOCK_GETTIME
 
 BOOST_REQUIRE([1.54])
-
-# Boost Context was introduced in 1.51 (Aug 2012), but there was an immediate
-# API break in 1.52 (Nov 2012), so we only support that, and later.
-pdns_context_library=""
-AS_IF([test $boost_major_version -ge 152], [BOOST_CONTEXT([], [no])])
-AS_IF([test x"$boost_cv_lib_context" = "xyes"], [
-  pdns_context_library="Boost Context"
-], [
-  AC_CHECK_FUNCS([getcontext makecontext swapcontext], [pdns_context_library="System V ucontexts"])
-])
-AC_MSG_CHECKING([what context library to use for MTasker])
-AS_IF([test -n "$pdns_context_library"], [AC_MSG_RESULT([$pdns_context_library])], [AC_MSG_ERROR([neither boost::context nor System V ucontexts available])])
+BOOST_CONTEXT([], [yes])]
 
 PDNS_ENABLE_UNIT_TESTS
 PDNS_ENABLE_REPRODUCIBLE
@@ -280,5 +269,4 @@ AS_IF([test "x$HAVE_LIBCURL" != "xn"],
   [AC_MSG_NOTICE([libcurl: yes])],
   [AC_MSG_NOTICE([libcurl: no])]
 )
-AC_MSG_NOTICE([Context library: $pdns_context_library])
 AC_MSG_NOTICE([])
